+++ src/main/kotlin/io/github/svioletg/positionPolling/DatabaseManager.kt
@@ -9,7 +9,24 @@ class DatabaseManager(private val plugin: PositionPolling) {
 
     fun setup() {
         val conn = this.getConnection()
-        val query = buildString {
+
+        conn.createStatement().execute(buildString {
+            append("CREATE TABLE IF NOT EXISTS plugin_info(")
+            append("version TEXT")
+            append(");")
+        })
+
+        // Ensure there's a single row, delete any exists ones and rebuild
+        conn.createStatement().execute("DELETE FROM plugin_info;")
+        conn.createStatement().execute(buildString {
+            append("INSERT INTO plugin_info(")
+            append("version")
+            append(") VALUES (")
+            append("'${this@DatabaseManager.plugin.meta.version}'")
+            append(");")
+        })
+
+        conn.createStatement().execute(buildString {
             append("CREATE TABLE IF NOT EXISTS player_positions(")
             append("timestamp REAL, ")
             append("player_uuid TEXT, ")
@@ -18,9 +35,7 @@ class DatabaseManager(private val plugin: PositionPolling) {
             append("y INTEGER, ")
             append("z INTEGER")
             append(");")
-        }
-
-        conn.createStatement().execute(query)
+        })
     }
 
     fun getConnection(): Connection {
@@ -58,5 +73,17 @@ class DatabaseManager(private val plugin: PositionPolling) {
 
             stmt.executeUpdate()
         }
+    }
+
+    fun entryCount(): Int {
+        val conn = this.getConnection()
+        var count = 0
+        val query = "SELECT * FROM player_positions;"
+        val result = conn.createStatement().executeQuery(query)
+        while (result.next()) {
+            count += 1
+        }
+
+        return count
     }
 }